Add tests that verify Studio can still render with bad HTML content.
